+++ src/main/java/com/NerdSpace/Rain/Entity/Projectile/TrainerProjectile.java
@@ -6,7 +6,7 @@ import com.NerdSpace.Rain.Graphics.Sprite;
 public class TrainerProjectile extends Projectile {
     public static final int RATE_OF_FIRE = 12;
 
-    public TrainerProjectile(int launchX, int launchY, double angle) {
+    public TrainerProjectile(double launchX, double launchY, double angle) {
         super(launchX, launchY, angle);
         sprite = Sprite.projectileTrainer;
         range = random.nextInt(75) + 100;
